Replace pre-built configure with autogen.sh
Best practice for autotools is to use autogen.sh to run autoreconf -i
which will generate configure.

The reason to not store configure in git is that it will change with
each revision of autoconf, which means each developer who runs autoreconf
will end up with a modified file in their git repository, which is not ideal.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>